requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.1
fastar==0.11.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.15
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.0
kagglehub==1.0.1
kagglesdk==0.1.23
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.5
packaging==26.2
pandas==2.2.3
pluggy==1.6.0
protobuf==7.34.1
pydantic==2.13.4
pydantic-extra-types==2.11.1
pydantic-settings==2.14.1
pydantic_core==2.46.4
Pygments==2.20.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.28
pytz==2026.2
PyYAML==6.0.2
requests==2.34.2
rich==15.0.0
rich-toolkit==0.19.9
rignore==0.7.6
scikit-learn==1.6.1
scipy==1.17.1
sentry-sdk==2.60.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uvicorn==0.47.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
